server/requirements.txt
pymupdf==1.25.5
Flask==3.1.0
ollama==0.4.8
Pillow==11.2.1
pytesseract==0.3.13
